rename ip_allocator_static to ip_allocator_pool
